requirements.txt
alabaster==0.7.12
attrs==19.3.0
audioread==2.1.8
autopep8==1.4.4
Babel==2.8.0
backcall==0.1.0
bleach==3.1.0
certifi==2020.6.20
cffi==1.13.2
chardet==3.0.4
cupy-cuda90==7.0.0
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
docutils==0.16
entrypoints==0.3
fastrlock==0.4
flake8==3.7.9
hydra-core==0.11.3
idna==2.10
imagesize==1.2.0
importlib-metadata==1.3.0
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.0
json5==0.8.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab==1.2.4
jupyterlab-server==1.0.6
jupyterthemes==0.20.0
kiwisolver==1.1.0
lesscpy==0.13.0
librosa==0.7.1
llvmlite==0.30.0
lxml==4.4.2
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mido==1.2.9
mistune==0.8.4
more-itertools==8.0.2
musdb==0.3.1
museval==0.2.0
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
numba==0.46.0
numpy==1.17.4
omegaconf==1.4.1
packaging==19.2
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.2
pathlib2==2.3.5
pexpect==4.7.0
pickleshare==0.7.5
pkg-resources==0.0.0
pluggy==0.13.1
ply==3.11
pretty-midi==0.2.9
prometheus-client==0.7.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
py==1.8.1
pyaml==19.12.0
pycodestyle==2.5.0
pycparser==2.19
pydub==0.23.1
pyflakes==2.1.1
Pygments==2.5.2
pyparsing==2.4.5
pyrsistent==0.15.6
pytest==5.3.2
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.2
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.24.0
resampy==0.2.2
scikit-learn==0.21.3
scipy==1.4.0
seaborn==0.9.1
Send2Trash==1.5.0
simplejson==3.17.0
six==1.13.0
sklearn==0.0
snowballstemmer==2.0.0
SoundFile==0.10.3.post1
Sphinx==3.2.1
sphinx-rtd-theme==0.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
stempeg==0.1.8
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
tqdm==4.40.0
traitlets==4.3.3
urllib3==1.25.10
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==0.6.0
